refactor: move components into updated file structure
moves row, navbar, avgbilled, coldstartsgraph components to new directories
