+++ client/src/components/Navbar/NavbarComponent.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import "../NavbarComponent.css";
-import lambda from "../assets/lambda.png";
+import "../../NavbarComponent.css";
+import lambda from "../../assets/lambda.png";
 
 const LightModeIcon = () => (
   <svg
