--- client/src/components/RowComponent.css
@@ -1,26 +1,0 @@
-.table-row {
-  display: grid;
-  grid-template-columns: 2fr 1fr 1fr 1fr;
-  gap: 10px;
-  padding: 12px 16px;
-  align-items: center;
-  background-color: #e1e1e1;
-  border-radius: 8px;
-  margin-bottom: 10px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.table-row:hover {
-  background-color: #f0f0f0;
-}
-
-body.darkmode .table-row {
-  color: #ffffff;
-  background-color: #363636;
-}
-
-body.darkmode .table-row:hover {
-  background-color: #a2a2a2;
-}
-
-/* All classes accounted for but need to import*/
