+++ client/src/containers/FunctionPerformance/AvgBilledDurGraph/AvgBilledDurGraph.css
@@ -1,0 +1,12 @@
+.chart-wrapper {
+  background-color: #e1e1e1;
+  border-radius: 8px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  padding: 16px;
+}
+
+body.darkmode .chart-wrapper {
+  background-color: #363636;
+}
+
+/* All classes accounted for but need to import*/
