+++ client/src/containers/FunctionPerformance/AvgBilledDurGraph/AvgBilledDurGraphComponent.tsx
