+++ client/src/containers/FunctionPerformance/ColdStartsMetrics/ColdStartsMetricsContainer.test.tsx
@@ -12,7 +12,7 @@ interface RowProps {
 // The RowComponent is mocked to simplify the tests and to isolate the container's behavior.
 // Instead of rendering the full implementation of RowComponent, it renders a simple div with a test id.
 // This allows the test to verify the container's logic without being coupled to the RowComponent's internals.
-jest.mock("../../../components/RowComponent", () => ({
+jest.mock("../../../components/Row/RowComponent.tsx", () => ({
   __esModule: true,
   default: (props: RowProps) => (
     <div data-testid="row">
